Convert save_to_cloud_storage job to sidekiq
The normal execution path for uploading to S3 via a worker is on the
transition of an assets state to "clean". This commit changes this
from a delayed-job worker to a sidekiq worker.

The existing delayed job worker (The code in
`Asset#save_to_cloud_storage`) cannot be removed at this stage, as it
is possible that there will still be jobs on the delayed job queue to
process when this code is deployed.

For now, we delegate the behaviour of Asset#save_to_cloud_storage to
SaveToCloudStorageWorker#perform. When there are no more jobs on the
delayed job queue we can remove this method.

In order to keep the spec moved from asset_spec to
save_to_cloud_storage_spec as similar as possible I've had to also
duplicate the two exceptions in `rubocop_todo` which this spec block
required. We can tackle these todo items later.