Redirect requests for unscanned Whitehall assets to placeholders
We've moved the `thumbnail-placeholder.png` image from Whitehall into
Asset Manager, because in the future we're imagining this functionality
being implemented for Mainstream assets too.

We've chosen not to move the HTML placeholder at
`/government/placeholder` for the moment, because it would involve
duplicating a lot of functionality from Whitehall for layout-related
content. We intend to revisit this decision at a later date.

This new behaviour is based on that in Whitehall's
`PublicUploadsController#fail` method. The main difference is that if
the asset is an image, we only redirect to the placeholder image if
the asset exists and is unscanned; whereas in Whitehall it was
sufficient for the asset to be an image - it didn't need to exist. I
belive the latter was an oversight - hence the difference here.

I've retained the existing behaviour for determining whether an image is
an image or not. However, I've captured #206 as a possibly more robust
approach.

Including `ActionView::Helpers::AssetUrlHelper` and using its `image_path`
method seems like a nicer way to generate the asset pipeline version of
the image URL path; as opposed to calling `view_context.path_to_image`
which is what Whitehall does.

The presence of an `app/assets` directoy causes the default Jenkins CI
build to run `bundle exec govuk-lint-sass app/assets/stylesheets`, but
since the latter directory doesn't exist, it would fail with a "No such
file or directory" exception. I considered adding the directory, but the
script still returns an error code if the directory is empty, so it
seems simpler to disable the SASS linting in the `Jenkinsfile` for now.

I haven't yet added any Cache-Control headers, but I plan to do that in
subsequent commits.