Add Cache-Control headers for Whitehall asset responses
Both these values match those used in Whitehall's
`PublicUploadsController`. I elected to store the main value in a new
configuration setting, `whitehall_cache_control`, because it was stored
in a configuration setting in Whitehall and it feels consistent with the
setting used in `MediaController#download`, i.e. `cache_control`.

One difference between the functionality in this commit compared to
that in Whitehall's `PublicUploadsController#fail` is that I've set the
short (1 min) cache expiry time for both placeholder redirects, not just
for non-image assets. I believe this was an oversight in Whitehall.

Note that we expect to bring these two values into line at some date in
the future, but for now it is simplest to keep the values unchanged.