Remove add_uuid_to_assets Rake task
I want to add validation to `Asset#uuid` to ensure that they're present,
unique, in the correct format and can't be changed after creation. The
presence of this data_migration_spec makes that hard because it relies
on the fact that some `Asset#uuid` values are nil.

This Rake task was added in b6480da246c3c01eccde507eefec54ac61df27c9 (as
part of PR #173) and has been run in integration, staging and
production. I've confirmed that all assets have a UUID in integration.
This data is copied from production so I'm confident that this Rake task
and associated code can be safely removed.