Validate format of Asset#uuid
This should ensure we don't end up with things that don't look like
UUIDs in the database. This shouldn't happen as long as we're
programmatically generating the UUIDs (using `SecureRandom.uuid` for
example) but will hopefully guard against us manually generating and
storing arbitrary values.

The regexp is Based on the "formal definition of the UUID string
representation" in RFC 4122[1].

Note that according to the RFC, "there is no mechanism for determining
whether a UUID is 'valid'".

[1]: https://tools.ietf.org/html/rfc4122