Move Asset#scan_for_viruses to Sidekiq worker
There is no change in behaviour if we move the implementation of this
method from Asset into a sidekiq worker and call it
synchronously (with `perform`). In a subsequent commit we'll call this
method asynchronously (on the sidekiq queue) but we'll need to leave
Asset#scan_for_viruses in place in this form in case there are still
any jobs on the delayed_job queue when this code is deployed.

We'll be able to remove this method later when there are no jobs on
the delayed_job queue.