Move spec of Asset#scan_for_viruses to worker spec
I've copied the specs for the method across pretty much as-is (only
changing the way the method is called in the specs). I've also added
the exceptions to rubocop that were required for this method in its
original form. To keep this commit simple, we can address those todo
items later.