Perform virus checking on sidekiq queue
This commit switches the asynchronous virus checking to use the
sidekiq worker instead of delayed_job. Due to the integration spec
`virus_scanning_spec` I've added a before spec hook to clear out the
sidekiq queues to avoid dependencies on the spec ordering.

The method `Asset#scan_for_viruses` becomes unreferenced with this
change, but as it may still be required if any delayed jobs are on the
queue when this is deployed, we'll leave it here for now. When we are
sure no new delayed jobs are being created it can be removed.