Introduce WhitehallAsset subclass and pull-down behaviour from Asset
I want to make the `legacy_url_path` mandatory when creating a Whitehall
asset. Explicitly modelling the extra state and behaviour provided by
such an asset will make it easier to introduce the relevant validation.

I've mostly been able to just *move* code from `Asset` into
`WhitehallAsset`, but I did have to make a few changes:

* `Asset#public_url_path` now only implements the "fallback" mainstream
  behaviour; while `WhitehallAsset#public_url_path` overrides it to
  implement the Whitehall-specific behaviour.

* Similarly, `Asset#mainstream?` now just returns `true` and no longer
  checks whether `legacy_url_path` is set, because it doesn't have
  access to the latter. `WhitehallAsset#mainstream?` just returns
  `false`.

I've also mostly been able to just *move* examples from the `Asset` spec
to the new `WhitehallAsset` spec, but I did make a few changes related
to the differences mentioned above.

I'm working on the assumption that no assets have been created with type
`Asset` and with `legacy_url_path` set. I did create one such asset in
integration when doing some testing, but it will have been overwritten
overnight when the production data was copied to integration.

Note that I've chosen to add the new `WhitehallAsset` spec to the
list of excluded files for the `RSpec/NestedGroups` Rubocop rule in the
to-do list, because its examples previously existed in the `Asset` spec
which was already excluded.