Remove Asset#scan_for_viruses
In 2ce4bfc we removed the last reference to this method. We left it
defined in case any jobs were on the delayed job queue that still
referenced it. With 2ce4bfc no new work is being added to delayed_job
so it is now safe to remove it.

To be very safe we should check that `Delayed::Job.count` is `0`
before deploying.