Remove Asset#save_to_cloud_storage
With 1bf0db64 we removed references to this method, but left it
defined in case any jobs on the delayed job queue still referenced it
at the time of deployment. With 1bf0db64 we are no longer adding work
to that queue so it is safe to remove this method.

To be very safe we should check that `Delayed::Job.count` is `0`
before deploying.