Add optional legacy_etag param when creating Whitehall asset
We're going to need this when we move existing Whitehall assets to Asset
Manager in order to avoid wholesale cache invalidation.

I've chosen to use `legacy_etag` vs just `etag` as the parameter name,
because once we've moved the existing Whitehall assets to Asset Manager
this will become redundant. Hopefully the name will also discourage
anyone from using the parameter in any of the other publishing apps.

I've aliased `WhitehallAsset#legacy_etag=` to `Asset#etag=` partly in
recognition that we'll only ever need to set `Asset#etag` via the API
when we move existing Whitehall assets to Asset Manager and partly to
keep the code as simple as possible. I think it's correct that the
canonical reader method is still just `Asset#etag` vs
`Asset#legacy_etag`, because that code is used more generally within
Asset Manager.