Set X-Frame-Options response header explicitly in Rails app
Currently the `X-Frame-Options` header is set to `DENY` for Mainstream
assets via the Nginx config for the `asset-manager` host [1]. For
Whitehall assets this header is set to the value in the response from
the Whitehall Rails app by the Nginx config for the `whitehall-admin`
host [2]. By default Rails sets this value to `SAMEORIGIN` [3,4] and the
default is not overridden in the Whitehall app config.

We're planning to serve Whitehall assets from Asset Manager. This commit
changes the Rails app to set the response header explicitly to "DENY"
for Mainstream assets and to "SAMEORIGIN" for Whitehall assets. Once
this change is deployed we will be able to change the Nginx config for
the `asset-manager` host to respect the value in the response from the
Asset Manager Rails app in order to preserve the externally visible
behaviour when we switch Whitehall assets to be served by Asset Manager.

In time, unless there's a good reason to use different values in
different scenarios, we should standardise on one of the values in all
scenarios.

[1]:
https://github.com/alphagov/govuk-puppet/blob/4cca939ec49a9b4c106b14b7cf896db31a003636/modules/govuk/manifests/apps/asset_manager.pp#L143-L151
[2]:
https://github.com/alphagov/govuk-puppet/blob/4cca939ec49a9b4c106b14b7cf896db31a003636/modules/govuk/manifests/apps/whitehall.pp#L235-L236
[3]:
https://github.com/rails/rails/blob/v4.2.7.1/actionpack/lib/action_dispatch/railtie.rb#L22
[4]:
http://guides.rubyonrails.org/v4.2.7.1/security.html#default-headers