Use Sidekiq to store file metadata on assets in Rake task
I ran into a problem running the Rake task on staging - the `each`
block was executed more times than the number of assets in the
database, i.e. the percent complete went well over 100%.

I suspect this was caused by a problem with the MongoDB cursor somehow
being reset within the `each` block, perhaps due to a timeout.

This commit changes the Rake task so it just queues up jobs to be
executed by a worker class. This way the execution of the `each` block
will be less than 1 min vs more than 15 mins. I'm hopeful that this will
avoid the cursor problem, because the original synchronous version of
the Rake task ran OK on staging despite taking nearly 15 mins to run and
I didn't see any evidence of the cursor problem.