Reduce priority of AssetFileMetadataWorker queue
We don't want AssetFileMetadataWorker jobs to interfere with
VirusScanWorker or SaveToCloudStorageWorker jobs.

This commit adds a new `low_priority` queue which will only be processed
when the `default` queue is empty, because the latter appears above it
in the list of queues in the configuration file.

Note that prior to this commit, the `default` queue existed implicitly
and it is the one that is used by default by the other worker classes.