Store WhitehallAsset#legacy_etag & #legacy_last_modified in database
This separates the concerns of (a) caching the attributes values
generated from the file metadata; and (b) overriding the attribute
values when migrating existing Whitehall assets to Asset Manager.

Note that we can now *always* overwrite the values of the attributes
within the `Asset#store_metadata` method, because the legacy values
are stored in the two new fields on `WhitehallAsset`.