Add spec for AssetFileMetadataWorker used by Rake task
When I first created the Rake task that uses this worker, the worker
didn't exist. Now that it is a separate class it's relatively easy to
test, so I thought I'd take the opportunity to add some test coverage.
The change is slightly orthogonal to other changes in this PR, but it's
still closely connected to the attributes on an asset which are
generated from file metadata, so I think it's OK to include it.