Make file metadata related attr setters on Asset protected
These attributes are only intended to be set from within the Asset
class. By making the attribute setter methods protected, we can make
this more obvious.

Note that the `govuk_assets:store_values_generated_from_file_metadata`
Rake task needs to set the attributes, but it is not affected by this
change, because AssetFileMetadataWorker#perform uses the Asset#set
method in order to bypass validation.