Remove ability to redirect asset requests to S3
We added this functionality in #112 when we were investigating the
performance of various approaches. Since we're not planning to use this
behaviour in the near future and I'm not confident that the behaviour is
as well tested as the other two alternatives (e.g. response headers), I
think it's better to remove the functionality for now - we can always
find this code in the git history if we need it again.

Note that I've retained `AssetManager.aws_s3_use_virtual_host`, because
although it was introduced in #112, it's now being used for
`S3Storage#presigned_url_for`.

Note also that I've checked that the `REDIRECT_ALL_ASSET_REQUESTS_TO_S3`
environment variable isn't beeing set anywhere in `govuk-puppet`.