Move examples to BaseMediaController spec
In order to test behaviour of the `proxy_to_s3_via_nginx?` method
in a more appropriate place. This will become more relevant when I
start re-using this method in `WhitehallMediaController#download`
which I plan to do in a subsequent commit.

Note that by moving these examples I've had to ignore the violations
of the RSpec/NestedGroups Rubocop rule which have moved to the new
`BaseMediaController` spec - hence the change to `.rubocop_todo.yml`.