Conditionally proxy requests for Whitehall assets to S3
Using the same logic we use for deciding whether to proxy requests for
Mainstream assets to S3 or serve them from NFS.

Note that at the time of making this change the
`PROXY_PERCENTAGE_OF_ASSET_REQUESTS_TO_S3_VIA_NGINX` environment
variable is set to 100 in production, but zero in staging and
integration. If that continues to be the case, it means that *all*
Whitehall asset requests will be proxied to S3 in production as soon as
these changes are deployed.

I can imagine it *might* be useful to be able to control the percentage
for Whitehall assets separately from that for Mainstream assets.
However, I'm going to leave that to a subsequent commit or possibly a
separate PR.