Add WhitehallAssetsController#show action
This allows us to find a WhitehallAsset using its `legacy_url_path`.
We're going to use this in Whitehall to obtain the ID of the asset so
that we can use the standard API to delete or update it.

Requests look something like
'/whitehall_assets/government/uploads/asset.png'. I can't use the
standard Rails restful route for the show action because that doesn't
appear to work for `:id` parameters that include slashes.

The implementation of `WhitehallAssets#show` is identical to
`Assets#show` in all but the way we find the asset.

I've chosen to assert that we're using the `AssetPresenter` to render
the asset (versus testing the actual rendered output) as the behaviour
of `AssetPresenter` is being unit-tested elsewhere.

I've added a single happy-path request spec to give us the some
confidence that we can change the implementation without breaking
things.