Independent control of proxying Whitehall vs Mainstream assets to S3
We're already proxying 100% of Mainstream asset requests to S3 in
production (although 0% in other environments).

These changes give us the flexibility to gradually ramp up proxying
Whitehall asset requests to S3 independent of the percentage set for
Mainstream asset requests.

I did contemplate encapsulating the differences in Mainstream vs
Whitehall asset configuration somewhere, but given that we're hoping
this is only going to be a temporary change, it didn't seem worthwhile.