Remove fallback from Asset attribute readers
Now that the `govuk_assets:store_values_generated_from_file_metadata`
Rake task has been run in all environments, we can rely on the presence
of the value in the database which is set on creation by
`Asset#store_metadata`.

I did consider adding validation for the presence of these attributes,
but it turned out to be very awkward and in any case, I realised that
it's a bit odd to validate attribtes which are set automatically by a
`before_save` callback, especially given that the attribute writer
methods are protected. So I decided that adding such validation was
unnecessary.