Don't rely on file metadata fallbacks in S3Storage spec
Prior to this change, the examples were relying on the fallbacks in the
file metadata related attribute reader methods on Asset. I'm about to
remove these fallbacks and so in order to keep this spec passing, I've
created the asset in the database which fires the callback to store the
file metadata related values.