Introduce S3Storage::Fake class
This implements a version of the S3Storage class backed by the local
file system rather than the AWS S3 service. It is intended to be used in
development environments, but I haven't wired it up yet, so it's not
currently being used.

Note that I've excluded `spec/lib/s3_storage/fake_spec.rb` from the
RSpec/PredicateMatcher Rubocop rule, because it was incorrectly
suggesting I could use the `exist` matcher.