Use S3Storage::Fake in dev environments if S3 not configured
We want to remove support for serving asset requests from NFS and proxy
all of them to S3. However, having to set up the necessary S3 resources
in a development environment seems rather onerous. The change in this
commit means that by default development environments will use a fake
version of S3Storage which stores the files in the local filesystem and
serves them via a Rack::File instance mounted in the Rails app.

Note that if the X-Accel-Mapping header is set in a request to the
Rack::File endpoint, this will trigger the Rack::Sendfile middleware
in the default Rails stack to send a response with the X-Accel-Redirect
header instead of the body of the file. This is currently a problem,
because the Nginx config for Asset Manager [1] sets the X-Accel-Mapping
request header whether or not the request is going to be served from NFS
or proxied to S3. Since the X-Accel-Mapping value is not set with the
fake S3 paths in mind, this results in an attempt to serve a
non-existent file and a 404 Not Found. We will need to address this
issue before merging this PR into master.

I struggled a bit to test for the presence of the Rack::File route, but
found that I could detect it if I gave it a name. Although this isn't
ideal, I think it's better than nothing.

[1]:
https://github.com/alphagov/govuk-puppet/blob/1b99dbc24b0ac477fadb75ef8d1356dbfd58bf2e/modules/govuk/manifests/apps/asset_manager.pp#L91-L92