Introduce S3Storage#set_metadata_for method
I want to manipulate the metadata on the S3 object for an asset in order
to trigger replication of objects that were in the S3 bucket *before*
replication to the backup bucket was enabled. Making these methods
available will allow me to do that in subsequent commits.

Note that calling Aws::S3::Object#put for a non-existent object seems to
add an object to the S3 bucket even if no body is specified. Given this
method is only intended to modifiy metadata on an existing S3 object, I
decided it would make sense to raise a
S3Storage::ObjectNotFoundError exception if the asset has no
corresponding object on S3.

Note that I'm happy not to implement the new S3Storage method on
S3Storage::Fake, because the method will only be used in a Rake
task that's only intended to be run on production.