Introduce S3Storage#exists? method
I want to iterate through all the assets and peform an operation on
their corresponding S3 object. However, there was no way to tell from
the model whether the upload to S3 had happened for a given asset.

Ideally, I think we would add a new state into the state machine, but
this will do for now.

I've disabled the RSpec/PredicateMatcher cop in the S3Storage spec,
because it doesn't make sense to use a predicate matcher in this case,
because the method under test takes an argument.