Simplify signature of methods in S3Storage::Null
There's not much value in duplicating the method signatures from
S3Storage and doing so for some methods I'm about to add would make the
dynamically generating suitable examples a lot more complicated, i.e.
we'd need to check the arity of each method and pass the correct number
of required arguments to each one.