Introduce S3Storage#add_metadata_to & #remove_metadata_from
I want to add and then remove a metadata key/value pair in order to
trigger replication of objects that were in the S3 bucket *before*
replication was enabled. Making these methods available will allow me to
do that in subsequent commits.

Note that these methods assume an S3 object exists for the asset and
will raise a S3Storage::ObjectNotFoundError if it doesn't. The idea is
that a client object should call S3Storage#exists? to decide whether
it's safe to call these methods.

Note that I'm happy not to implement the new S3Storage methods on
S3Storage::Fake, because the methods will only be used in a Rake
task that's only intended to be run on production.

Also note that I've introduced an `exclude` RSpec matcher in
`spec/support/matchers/exclude.rb`.