Add Rake task to trigger replication for all assets
S3 objects for some assets were added to the S3 bucket *before*
replication was enabled on the bucket. Enabling replication does not
automatically replicate these S3 objects, so we need to do this
manually.

The simplest way to achieve this seems to be to modifiy the metadata
on the S3 object. This feels better than manually syncing/copying the
relevant S3 objects to the target bucket, because it uses the same
mechanism (replication) to sync to the target bucket with all the same
users, roles & permissions.

One possible downside of this approach is that the Last-Modified
timestamp updated when the metadata on an S3 object is modified.
However, at the moment, we're not making use of this metadata - the
Rails app sets a Last-Modified response header using a value stored in
the database.

Another possible downside is that the new metadata entry will be sent as
a response header when the S3 object is requested. However, the worker
immediately removes the new metadata entry, so this will only ever be a
very transient issue and in any case, the extra response header
shouldn't cause any problems.

Note also that I've extracted a method to remove the duplication between
the new Rake task and the existing Rake task and to avoid a Rubocop
violation due to the length of the block withint the Rake namespace!