Only trigger replication for non-replicated S3 objects in Rake task
Previously the AssetTriggerReplicationWorker updated metadata on the
S3 objects for all assets in order to trigger replication whether or
not they had already been replicated. Now the worker only updates the
metadata on S3 objects which have no replication status - this is the
status that S3 objects have if they are added to an S3 bucket before
replication is enabled.

Note that the replication status can take the following values:

* nil
* 'PENDING'
* 'COMPLETED'
* 'FAILED'
* 'REPLICA'

Note that I'm happy not to implement the new S3Storage method on
S3Storage::Fake, because the method is only currently used in a Rake
task that's only intended to be run on production.