Introduce S3Storage#never_replicated? method
I want to identify S3 objects which have never been replicated so that I
can limit the number of S3 object for which I have to manually trigger
replication via the govuk_assets Rake task.

Note that the replication status can take the following values:

* nil
* 'PENDING'
* 'COMPLETED'
* 'FAILED'
* 'REPLICA'

Note that I'm happy not to implement the new S3Storage method on
S3Storage::Fake, because the method is only going to be used in a
Rake task that's only intended to be run on production.