Only trigger replication for non-replicated S3 objects in Rake task
Previously the AssetTriggerReplicationWorker re-saved the S3 object
for all assets in order to trigger replication whether or not they
had already been replicated. Now the worker only re-saves the S3
objects which have a nil replication status - this is the status that
S3 objects have if they are added to an S3 bucket before replication
is enabled.

I've also renamed the Rake task and changed its description to reflect
these changes.