Generate correct path to thumbnail-placeholder.png
We redirect to this placeholder in response to requests for unscanned
Whitehall images.

We were incorrectly redirecting to '/images/thumbnail-placeholder.png'
instead of '/asset-manager/thumbnail-placeholder-<digest>.png'. This
appears to be because the `image_path` method from
`ActionView::Helpers::AssetUrlHelper` doesn't take the Asset Pipeline
into account. The code snippets below illustrate the problem and fix:

    # This ignores our assets.prefix config option
    >> Class.new do
         include ActionView::Helpers::AssetUrlHelper
       end.new.image_path('thumbnail-placeholder.png')
    => "/images/thumbnail-placeholder.png"

    # This respects our assets.prefix config option
    >> ActionController::Base.helpers.image_path('thumbnail-placeholder.png')
    => "/asset-manager/thumbnail-placeholder-b0b645c0f44c2fc7d8cc2fee24f576d82256a85ac3c3e0a59d6ff60d2903d6b6.png"

I've confirmed that redirecting to the thumbnail containing a digest is
the same as the behaviour in Whitehall (which is what we're interested
in replicating):

    $ curl -I https://assets.publishing.service.gov.uk/government/uploads/foo.png \
      | grep Location

    Location: //whitehall-admin.publishing.service.gov.uk/government/assets/thumbnail-placeholder-b0b645c0f44c2fc7d8cc2fee24f576d82256a85ac3c3e0a59d6ff60d2903d6b6.png