Allow arbitrary scopes for AssetProcessor
I want to be able to process only deleted assets in a Rake task and this
will make that easier.

Note that I had to move the construction of the AssetProcessor into the
individual Rake tasks where the Rails environment is available in order
to allow the use of Asset in the AssetProcessor constructor arguments
without causing an exception when Rake is run and the
`lib/tasks/govuk_assets.rake` file is loaded.