Add Rake task to upload deleted assets to S3
We recently realised [1] that when we did the initial upload of existing
assets to S3, we forgot to include assets marked as deleted. Since these
assets can feasibly be restored, we ought to store the relevant file on
S3 like any other asset.

Note that uploading the files to S3 will also have the effect of
triggering replication which was triggered for non-deleted assets using
the `trigger_replication_for_non_replicated_assets` Rake task.

Note also that I had to disable the `Metrics/BlockLength` Rubocop rule
around the Rake namespace, because it had got too long. Initially I
tried disabling this cop for the whole file in the Rubocop todo config,
but it seemed to negate the default config which disables the rule for
all specs. It doesn't seem so important to keep a namespace block short,
so I'm happy to do this.

[1]: https://github.com/alphagov/asset-manager/issues/301