Add worker to store file metadata values for deleted asset
I'm about to add a new Rake task to store attributes generated from file
metadata in the database for assets which have been marked as deleted. I
contemplated re-using the existing AssetFileMetadataWorker, but given
that all of the "low priority" workers are associated with one-off Rake
tasks, I decided to add a new worker instead.