Add Rake task to permanently delete asset
The app uses the mongoid_paranoia gem which means that assets deleted
via the API are only marked as deleted. This is so they can be restored
via the API.

Sometimes it's desirable to permanently remove an asset, e.g. if an
asset containing sensitive information is accidentally uploaded. This
Rake task allows us to permanently delete an asset which has already
been marked as deleted.