Set AWS_S3_BUCKET_NAME to a placeholder on CI
In CI we want to run asset compilation in production mode to catch any
errors before we deploy as there are differences between dev/test mode
compilation and production mode.  This app requires that
AWS_S3_BUCKET_NAME is set in production, so we set that envvar in the
Jenkinsfile after we've run the tests (asset compilation is done after
the other tests).