Make feature flag settings explicit in request specs
Some of the examples in these specs were relying on the fact that the
environment variables corresponding to the relevant feature flags were
not set and hence that the values were defaulting to 0.

By stubbing the feature flag methods we can make it clearer the context
in which the examples are intended to run.