Extract let statements in whitehall media requests spec
I want to refer to these assets in a subsequent commit. Making this
change first will make the diff for the subsequent change easier to
understand.

I've had to use `let!` vs `let` for the moment, because they're not yet
being referenced. I plan to change this in the subsequent commit.