Make virus scanning spec more realistic
Currently the feature flag is set to 100 in all environments [1], so
this makes the spec more realistic.

I've had to remove the assertion that the response body matches the
content of the asset fixture file, because now the context of the spec
is that the app is proxying asset requests to S3 via Nginx and so the
body of the response from the Rails app is now empty. I'm happy to do
this, because I think the assertion is somewhat orthogonal to the
motivation behind the example and I'm confident the app's behaviour to
instruct Nginx to proxy to S3 is well tested elsewhere, notably in the
two media request specs.

[1]:
https://github.com/alphagov/govuk-puppet/blob/476d9e764b42348deb8a2253be78d460d44c33c0/hieradata/common.yaml#L523