Remove request params for forcing proxying to S3 via Nginx
Currently both feature flags are set to 100 in all environments [1], so
all asset requests are being proxied to S3 via Nginx. That means that
there's no longer any point in having these request parameters to
override the settings set in the environment variables, especially as
the latter will probably be removed shortly.

[1]:
https://github.com/alphagov/govuk-puppet/blob/476d9e764b42348deb8a2253be78d460d44c33c0/hieradata/common.yaml#L523-L524