Remove out-of-date documentation
We haven't updated this documentation in ages. While it contains useful
content, I think it gives a misleading impression of the state of the
project and so I've moved it into the "Asset Management" folder in
the GDS Google Drive. The GitHub milestones [1], issues and pull
requests on the asset-manager repo are a better representation of the
state of the project.

[1]: https://github.com/alphagov/asset-manager/milestones