Hard-code app to proxy all asset requests to S3 via Nginx
Currently both feature flags are set to 100% in all environments [1], so
all asset requests are being proxied to S3 via Nginx. We're happy that
this mechanism is working OK and have no plans to rollback to serving
assets from NFS via Nginx, so there's not really any point in the app
taking notice of these environment variables any more.

Doing this should mean that we can simplify the application in
subsequent commits.

Note that I've also removed the stubbing of the two feature flag
attributes, because they are now hard-coded to the return values from
the stubbed methods.

[1]:
https://github.com/alphagov/govuk-puppet/blob/476d9e764b42348deb8a2253be78d460d44c33c0/hieradata/common.yaml#L523-L524