Hard-code BaseMediaController#proxy_to_s3_via_nginx? to true
Currently both feature flags are set to 100% in all environments [1], so
all asset requests are being proxied to S3 via Nginx. We're happy that
this mechanism is working OK and have no plans to rollback to proxying
less than 100% of asset requests to S3 via Nginx, so there's no longer
any need for the random number generator in the
BaseMediaController#proxy_to_s3_via_nginx? method.

Doing this should mean that we can simplify the application in
subsequent commits.

[1]:
https://github.com/alphagov/govuk-puppet/blob/476d9e764b42348deb8a2253be78d460d44c33c0/hieradata/common.yaml#L523-L524