Make GOV.UK app host configuration more flexible
I want to be able to set the app host to "http://localhost" or
"http://127.0.0.1" for the publishing-e2e-tests project. Previously I
could bodge this by setting GOVUK_APP_NAME to "127" and GOVUK_APP_DOMAIN
to "0.0.1", but making this change means we now have the flexiblity to
just set GOVUK_APP_DOMAIN to "localhost" or "127.0.0.1".