POST /whitehall_assets replaces existing asset
Currently in Whitehall images stored using the `ImageUploader` are
stored on NFS only. We want to change Whitehall to store these images
in Asset Manager.

In some circumstances an image using this Uploader (such as the image
associated with a Person[1]) can be replaced. For example

1. A new person is created with an image `foo.jpg`
2. The person is saved
3. The person is edited to have a different image with the same
filename `foo.jpg`
4. The person is saved again

In this sequence of events the new image will be available at the same
URL as the image added in step 1. The image added in step 1. will be
deleted from the file system.

In order to allow this behaviour in Asset Manager this commit modifies
`POST /whitehall_assets` to check whether an asset with the same
`legacy_url_path` already exists and soft-deletes it before creating
the new asset.

This will allow us to maintain the desirable existing behaviour from
Whitehall (a new image served at an existing URL) and also allows the
replaced image to be recovered if required[2].

[1] /government/admin/people/ in Whitehall Admin