Fix Whitehall asset lookup when URL path has no format/file ext
We were re-constructing the `legacy_url_path` to search for by assuming there
was a format [1]. Thus if an asset had no file extension, e.g.
`/government/uploads/foo`, we were attempting to lookup the asset using
`/government/uploads/foo.` (note the trailing full-stop). This did not
find the asset and the app responded with a `404 Not Found`.

[1]:
https://github.com/alphagov/asset-manager/blob/c14c324b4f41467ab6356599bc95ea92b34b99f6/app/controllers/whitehall_media_controller.rb#L3-L4