Avoid unnecessary asset download from S3
This should avoid the problem captured in issue #238 ("Binary file
content being displaying in Internet Explorer").

We believe that the problem was being triggered when Fastly made a
conditional GET request for an asset that it had stored in its cache.
Rails (or more specifically, `Rack::ConditionalGet`[1]) would respond
with a 304 Not Modified and an empty `Content-Type` and empty
`Content-Length` header. This would've been fine if it didn't also send
the `X-Accel-Redirect` header. The presence of this header was telling
Nginx to request the file from S3 again. Nginx would then send a 200
response containing the asset but with an empty `Content-Type` header!

By not sending the `X-Accel-Redirect` header along with a 304 response,
we ensure that Nginx passes the 304 back to Fastly as expected.

[1]:
https://github.com/rack/rack/blob/master/lib/rack/conditional_get.rb