Remove option to use S3 virtual host for asset URLs
This was added for experimental use in development only [1] and has
never been enabled in any other environments. Although eventually we
might end up wanting to use this functionality, at the moment it's
just making the code more complicated for no good reason.

Note that I've left the redundant S3Storage spec context in place to
make it easier to see what's going on in the diff, but I plan to remove
it in a subsequent commit.

[1]: 552e396ae5200c38852674b7dafcdc65467b5cb5