Extract aws_s3_bucket_name into S3Configuration class
This makes it easier to test and will make it easier to add related
options in subsequent commits.

Note that this means that the app no longer fails fast on startup if the
S3 bucket is not configured in a Rails production environment. I plan to
fix this in a later commit.