Increase Max option for RSpec/NestedGroups cop from 2 -> 4
I was triggering violations of this cop quite regularly. I think the
default value of 2 is a bit draconian and rather than continuing to add
files to the exclusion list, I think it's more sensible to bump up the
value to 4.

Meanwhile I've opened an issue [1] to suggest this should become the
default for govuk-lint.

[1]: https://github.com/alphagov/govuk-lint/issues/91