Extract S3Configuration#configured? method
We're using the presence of the S3 bucket name in the configuration as a
short-cut way to decide whether the S3 bucket is configured. In reality
other things, notably the AWS credentials, will need to be configured.
By extracting this method and adding specs, we can make this intention
clearer.