Use S3Configuration#fake? to decide whether to setup fake S3 route
This reduces duplication of the logic and ensures the route is setup
when it is needed and not setup when it is not needed.

Note that this indirectly means that the app will fail fast on startup
if the S3 bucket is not configured in a Rails production environment.
However, I plan to make this behaviour more explicit in a later commit.