Remove S3Storage::Null which is no longer needed
This class was introduced in this commit [1] so that we could deploy the
application before the relevant environment variables were setup using
govuk-puppet. At that point we were not using the S3 bucket in anger and
we wanted to avoid a stream of exceptions occurring when the
SaveToCloudStorageWorker couldn't save any new assets because S3 was not
properly configured.

Now the S3 bucket is being used in earnest it no longer makes sense to
have this as a fallback - in fact it might be a bit dangerous - so it
seems better and simpler to fail fast if S3 is not configured and we
can't use the fake S3 implementation.

[1]: 1447c0d77ea33ec22ba45cc0ba52abd6e0baf2cd